Internal function for calculating data for funnel plot

This is simplified version of DALEXtra::funnel_measure

funnel_measure(
  explainer,
  measure_function = NULL,
  nbins = 5,
  cutoff = 0.01,
  cutoff_name = "Other",
  factor_conversion_threshold = 7
)

Arguments

explainer
Explainer created using DALEX::explain

measure_function
measure function that calculates performance of model based on true observation and prediction. Order of parameters is important and should be (y, y_hat). The measure calculated by the function should have the property that lower score value indicates better model. If NULL, RMSE will be used for regression, one minus auc for classification and crossentropy for multiclass classification.

nbins
Number of qunatiles (partition points) for numeric columns. In case when more than one qunatile have the same value, there will be less partition points.

cutoff
Threshold for categorical data. Entries less frequent than specified value will be merged into one category.

cutoff_name
Name for new category that arised after merging entries less frequent than cutoff

factor_conversion_threshold
Numeric columns with lower number of unique values than value of this parameter will be treated as factors

Value

Data frame with columns

Variable Name of splited variable

Measure Loss value for subset

Label Label for variable's values subset